add waste electricity demand to fuels sector for hydrogen reconversion
